resvisi setelah sidang p1

+++ app/Http/Controllers/Reservation/ReservationGuestController.php
@@ -32,7 +32,10 @@ class ReservationGuestController extends Controller
     public function create()
     {
         $rooms = Room::all();
-        return view('frontOffice.reservation.create',  \compact('rooms'));
+        $roomRateStandart = Room::where('roomType', 'standart')->first();
+        $roomRateSuperior = Room::where('roomType', 'superior')->first();
+        $roomRateDeluxe = Room::where('roomType', 'deluxe')->first();
+        return view('frontOffice.reservation.create',  \compact('rooms', 'roomRateStandart', 'roomRateSuperior', 'roomRateDeluxe'));
     }
 
     /**
@@ -40,8 +43,9 @@ class ReservationGuestController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
-     */
-    public function store(ReservationFormRequest $request)
+     */ 
+    //ReservationFormRequest
+    public function store(Request $request)
     {
         //code mengambil selisih antara checkin dan checkout
         $checkIn = new Carbon($request->arrivaleDate);
@@ -49,18 +53,23 @@ class ReservationGuestController extends Controller
         $difference = ($checkIn->diff($checkOut)->days < 1)
             ? 'today'
             : $checkIn->diffInDays($checkOut);
-        $data = \array_merge($request->except('_token', 'rooms', 'totalRoomReserved', 'roomRate'));
+        $data = \array_merge($request->except('_token', 'rooms', 'totalRoomReserved', 'roomRate', 'discount'));
         $reservation = Reservation::create($data);
 
+        //menyimpan pengaturan kamar reservasi
         if (\count($request->rooms) > 0) {
             foreach ($request->rooms as $room => $v) {
-                $individualReservationDetail = [
-                    'reservation_id' => $reservation->id,
-                    'totalRoomReserved' => $request->totalRoomReserved[$room],
-                    'typeOfRoom' => $request->rooms[$room],
-                    'roomRate' => $request->roomRate[$room]
-                ];
-                DB::table('individual_reservation_rooms')->insert($individualReservationDetail);
+                //jika req room ada dan req totalRoomReserved ada, maka true
+                if ($request->rooms[$room] != '' && $request->totalRoomReserved[$room]  != '') {
+                    $individualReservationDetail = [
+                        'reservation_id' => $reservation->id,
+                        'totalRoomReserved' => $request->totalRoomReserved[$room],
+                        'typeOfRoom' => $request->rooms[$room],
+                        'roomRate' => $request->roomRate[$room],
+                        'discount' => $request->discount[$room]
+                    ];
+                    DB::table('individual_reservation_rooms')->insert($individualReservationDetail);
+                }
             }
         }
         
@@ -83,18 +92,7 @@ class ReservationGuestController extends Controller
         $difference = ($checkIn->diff($checkOut)->days < -1)
             ? 'today'
             : $checkIn->diffInDays($checkOut);
-        $totalRate = $reservation->individualReservationRooms()->sum('roomRate') * $difference;
-
-        $individualReservationRoom = DB::table('individual_reservation_rooms')
-                                ->where('reservation_id', $reservation->id)
-                                ->get();
-
-        $total = 0;
-        foreach($individualReservationRoom as $value){
-            $total += $value->roomRate * $value->totalRoomReserved * $difference;
-        }
-
-        return view('frontOffice.reservation.detail', \compact('reservation', 'difference', 'total'));
+        return view('frontOffice.reservation.detail', \compact('reservation'));
     }
 
     /**
@@ -111,6 +109,7 @@ class ReservationGuestController extends Controller
                 ->where('typeOfRoom', '=', 'extraBad')
                 ->where('reservation_id', '=' ,$id)
                 ->first();
+
         return view('frontOffice.reservation.edit', \compact('reservation', 'rooms', 'extraBad'));
     }
 
@@ -124,27 +123,30 @@ class ReservationGuestController extends Controller
     public function update(Request $request, Reservation $reservation)
     {
         // \dd($request->all());
-        $data = $request->except('_token', 'numberRoom', 'totalRoomReserved', 'rooms', 'idRooms', 'roomRate');
+        $data = $request->except('_token', 'numberRoom', 'totalRoomReserved', 'rooms', 'idRooms', 'roomRate', 'discount');
         $reservation->update($data);
 
         //ROOM ARRAGEMENT
-        for ($i = 0; $i < \count($request->idRooms); $i++) {
+        for ($i = 0; $i < \count(array($request->idRooms)); $i++) {
             DB::table('individual_reservation_rooms')->where('id', $request->idRooms[$i])
                 ->update([
                     'reservation_id' => $reservation->id,
                     'totalRoomReserved' => $request->totalRoomReserved[$i],
                     'typeOfRoom' => $request->rooms[$i],
-                    'roomRate' => $request->roomRate[$i]
+                    'roomRate' => $request->roomRate[$i],
+                    'discount' => $request->discount[$i]
                 ]);
         }
         //jika ada tambahan kamar
+        // jika request room lebih besar dari pada data kamar sebelumnya, maka ada tambahan kamar
         if (\count($request->rooms) > \count($reservation->individualReservationRooms)) {
             foreach ($request->rooms as $room => $v) {
                 $individualReservationDetail = [
                     'reservation_id' => $reservation->id,
                     'totalRoomReserved' => $request->totalRoomReserved[$room],
                     'typeOfRoom' => $request->rooms[$room],
-                    'roomRate' => $request->roomRate[$room]
+                    'roomRate' => $request->roomRate[$room],
+                    'discount' => $request->discount[$room]
                 ];
                 IndividualReservationRoom::updateOrCreate($individualReservationDetail, [
                     'reservation_id' => $reservation->id
@@ -163,9 +165,21 @@ class ReservationGuestController extends Controller
      */
     public function destroy(Reservation $reservation)
     {
-        $reservation->update(['status' => 0]);
-        $reservation->delete();
-        return \response()->json(['sukses' => true]);
+        if ($reservation->status == 'checkIn') {
+            $success = true;
+            $message = 'Guest has chacked in, data cannot be deleted';
+        } else {
+            $reservation->update(['status' => 0]);
+            $reservation->delete();
+
+            $success = false;
+            $message = 'Data has been deleted';
+        }
+
+        return \response()->json([
+            'success' => $success,
+            'message' => $message
+        ]);
     }
 
     public function cancelGuest()
@@ -193,5 +207,171 @@ class ReservationGuestController extends Controller
         $individualReservationDetail = IndividualReservationRoom::find($id);
         $individualReservationDetail->delete();
         return \redirect()->back();
+    }
+
+    /**
+     * Jquery
+     */
+
+    public function searchRoomRate(Request $request)
+    {
+        $roomRate = DB::table('rooms')->where('roomType', $request->typeRoom)->first();
+        return \response()->json([
+            // 'success' => $success,
+            // 'message' => $message,
+            'room' => $roomRate,
+        ]);
+    }
+
+    public function checkRoomStandart(Request $request)
+    {
+        $totalRoomReservedStandart = DB::table('individual_reservation_rooms')->where([
+            ['typeOfRoom', '=', 'standart'],
+            ['status', '=', 1]
+        ])->sum('totalRoomReserved');
+
+        $roomStandartVR = Room::where([
+            ['roomType', '=', 'standart'],
+            ['code' ,'=', 'VR']
+        ])->get();
+
+        if ($request->totalRoomReserved > \count($roomStandartVR) - $totalRoomReservedStandart) {
+            $success = true;
+            $message = 'The room STANDART booked exceeds the available rooms';
+        } else {
+            $success = \false;
+            $message = $request->totalRoomReserved.' Rooms standart available';
+        }
+
+        return \response()->json([
+            'success' => $success,
+            'message' => $message,
+        ]);
+    }
+
+    public function checkRoomSuperior(Request $request)
+    {
+        $roomSuperiorVR = Room::where([
+            ['roomType', '=', 'superior'],
+            ['code', '=', 'VR']
+        ])->get();
+        $totalRoomReservedSuperior = DB::table('individual_reservation_rooms')->where([
+            ['typeOfRoom', '=', 'superior'],
+            ['status', '=', 1]
+        ])->sum('totalRoomReserved');
+
+        if ($request->totalRoomReserved > \count($roomSuperiorVR) - $totalRoomReservedSuperior) {
+            $success = true;
+            $message = 'The room SUPERIOR booked exceeds the available rooms';
+        } else {
+            $success = \false;
+            $message = $request->totalRoomReserved.' Rooms superior available';
+        }
+
+        return \response()->json([
+            'success' => $success,
+            'message' => $message,
+        ]);
+    }
+
+    public function checkRoomDeluxe(Request $request)
+    {
+        $totalRoomReservedDeluxe = DB::table('individual_reservation_rooms')->where([
+            ['typeOfRoom', '=', 'deluxe'],
+            ['status', '=', 1]
+        ])->sum('totalRoomReserved');
+
+        $roomDeluxeVR = Room::where([
+            ['roomType', '=', 'deluxe'],
+            ['code', '=', 'VR']
+        ])->get();
+
+        if ($request->totalRoomReserved > count($roomDeluxeVR) - $totalRoomReservedDeluxe) {
+            $success = true;
+            $message = 'The room DELUXE booked exceeds the available rooms';
+        } else {
+            $success = \false;
+            $message = $request->totalRoomReserved.' Rooms deluxe available';
+        }
+
+        return \response()->json([
+            'success' => $success,
+            'message' => $message,
+        ]);
+    }
+
+    public function checkAvailableRoom(Request $request)
+    {
+        //STANDART
+        $totalRoomReservedStandart = DB::table('individual_reservation_rooms')->where([
+            ['typeOfRoom', '=', 'standart'],
+            ['status', '=', 1]
+        ])->sum('totalRoomReserved');
+
+        $roomStandartVR = Room::where([
+            ['roomType', '=', 'standart'],
+            ['code' ,'=', 'VR']
+        ])->get();
+
+        //SUPERIOR
+        $roomSuperiorVR = Room::where([
+            ['roomType', '=', 'superior'],
+            ['code', '=', 'VR']
+        ])->get();
+        $totalRoomReservedSuperior = DB::table('individual_reservation_rooms')->where([
+            ['typeOfRoom', '=', 'superior'],
+            ['status', '=', 1]
+        ])->sum('totalRoomReserved');
+
+        //DELUXE
+        $totalRoomReservedDeluxe = DB::table('individual_reservation_rooms')->where([
+            ['typeOfRoom', '=', 'deluxe'],
+            ['status', '=', 1]
+        ])->sum('totalRoomReserved');
+
+        $roomDeluxeVR = Room::where([
+            ['roomType', '=', 'deluxe'],
+            ['code', '=', 'VR']
+        ])->get();
+
+        if($request->typeRoom == 'standart'){
+
+            if ($request->totalRoomReserved > \count($roomStandartVR) - $totalRoomReservedStandart) {
+                $success = true;
+                $message = 'The room STANDART booked exceeds the available rooms';
+            } else {
+                $success = \false;
+                $message = $request->totalRoomReserved.' Rooms standart available';
+            }
+        } elseif($request->typeRoom == 'superior'){
+
+            if ($request->totalRoomReserved > \count($roomSuperiorVR) - $totalRoomReservedSuperior) {
+                $success = true;
+                $message = 'The room SUPERIOR booked exceeds the available rooms';
+            } else {
+                $success = \false;
+                $message = $request->totalRoomReserved.' Rooms superior available';
+            }
+
+        } elseif($request->typeRoom == 'deluxe'){
+            
+            if ($request->totalRoomReserved > count($roomDeluxeVR) - $totalRoomReservedDeluxe) {
+                $success = true;
+                $message = 'The room DELUXE booked exceeds the available rooms';
+            } else {
+                $success = \false;
+                $message = $request->totalRoomReserved.' Rooms deluxe available';
+            }
+
+        } else {
+            $success = \false;
+            $message = 'Total room reserved or type room is not blank';
+        }
+
+
+        return \response()->json([
+            'success' => $success,
+            'message' => $message,
+        ]);
     }
 }
